Consolidate regular and collection uploader
