+++ client/src/components/Upload/utils.js
@@ -11,6 +11,11 @@ export const AUTO_EXTENSION = {
     description:
         "This system will try to detect the file type automatically. If your file is not detected properly as one of the known formats, it most likely means that it has some format problems (e.g., different number of columns on different rows). You can still coerce the system to set your data to the format you think it should be.  You can also upload compressed files, which will automatically be decompressed.",
 };
+export const COLLECTION_TYPES = [
+    { id: "list", text: "List" },
+    { id: "paired", text: "Pair" },
+    { id: "list:paired", text: "List of Pairs" },
+];
 export const DEFAULT_DBKEY = "?";
 export const DEFAULT_EXTENSION = "auto";
 export const DEFAULT_FILE_NAME = "New File";
